working copy

+++ way.h
@@ -22,7 +22,7 @@ public:
         way(way* prev,way* next);
         int read(way** head, ull address);
         int write(way** head, ull address);
-        int fill(way** head, ull address);
+        ull fill(way** head, ull address);
         int flush(way** head);//flush all ways
 };
